requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
email-validator>=2.0.0
SQLAlchemy==2.0.36
oss2==2.19.1
httpx==0.27.2
openai==1.57.0
dashscope==1.20.11
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-multipart==0.0.12


